pyproject.toml
[tool.poetry]
name = "turing-machine"
version = "0.1.0"
description = "Simple GUI for Turing Machine"
authors = ["Pavel Barabanov <barabanovpv95@gmail.com>"]
readme = "README.md"
packages = [{include = "turing_machine"}]

[tool.poetry.dependencies]
python = "^3.11"
pyqt6-tools = "^6.4.2.3.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
